*** empty log message ***
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@126 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/vizor/kernel/v_user.h
@@ -154,7 +154,7 @@ unsigned int lenWithRightStrip(const char* s)
 
 #ifdef __FreeBSD
 // replacement for missing function in FreeBSD
-inline char* gcvt(double num, size_t digit, char* buf)
+inline char* gcvt(double num, int digit, char* buf)
 {
     sprintf(buf, "%*g", digit, num);
     return buf;
